Added Resupply Truck
Adds a "Resupply Truck" to vehicle Store. This truck can be used to
repair, refuel and rearm. This operations take some time. The player is
shown what's happening in his vehiclechat. I look up the display name of
magazines if set. Otherwise classnames are used. Was tested to be
working for all kinds of vehicles and static weapons.