Make hunger, and thirst 100 if not avaiilable
If player disconnects, and for some reason hunger, and thirst
are not available at least give a good default, of 100.

This can happen if their game crashes, before they have reported hunger
and thirst.

These two values are only reported at fixed intervals, or when the player
presses ESC. This is in order to avoid spamming the server because these
values change very often.