Fix regression with beacons, and warchest saving
After the peformance fix for object saving ... beacons were not
in the "locked_objects_list" anymore, and were not being saved. I have
also nose refactored the code to rename the list to
"tracked_objects_list" so that it makes more sense, it could also contain
beacons, and warchests which are not necessarily locked/unlocked.