Save active players list in database
Save the current list of players in the database,
and also save their basic information, as well as
their score. This should make it really simple
to display a pseudo-live scoreboard externally on a
website.

There is a new config variable: A3W_playersList_saveInterval
which controls how often the player's list is saved to the
database. By default it's every 2 minutes.