Add multi-map stats support (saved position)
Make position be saved based on the map.

Position_Altis
Position_Stratis

Still support legacy "Position" for Altis.
If "Position_Altis" is not available, but "Position" is available, use it.

Also, if there is no position available at all for any of the maps, then
put the player in a random safe position (so that he does not stay in
spawn island)

Conflicts:
	persistence/players/sFunctions.sqf