Modify order in which weapon items are restored
Sometimes when player has a lot of gear, and weapon items are restored
last ... the items do not get added to the weapon. This causes grief for
players who buy expensive items like the night stalker scope.

This fix makes it so that weapon items are added as early as possible.

Also changed some logging calls so that the RPT shows all data that came
from the database for player information.